requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
motor==3.3.2
minio==7.2.3
redis==5.0.1
python-dotenv==1.0.1
websockets==12.0